Refactor files in model
File names did not reflect content, so updated best I could to make it correct.
