+++ model/hardware.go
